자원 예약 수정
