+++ http/mailRestAPI.http
@@ -4,7 +4,7 @@ Content-Type: multipart/form-data; boundary=WebAppBoundary
 Authorization: Bearer eyJkYXRlIjoxNzE2MTc0MTU3NTIwLCJ0eXBlIjoiand0IiwiYWxnIjoiSFMyNTYifQ.eyJwb3NpdGlvbk5hbWUiOiLtjIDsnqUiLCJkZXBObyI6MiwiaW1nVXJsIjoiYzExNDc3MzJkNmMxNDk2OThmOGI0ZGM2NjVmMzlmZTcucG5nIiwibWVtYmVyTm8iOiIyIiwicG9zaXRpb25MZXZlbCI6Miwic3ViIjoia2V0Y2h1cCB0b2tlbiA6IDIiLCJyb2xlIjoiTFYyIiwicG9zaXRpb25TdGF0dXMiOiJZIiwibWVtYmVyTmFtZSI6IuuwleuLpOydgCIsInBvc2l0aW9uTm8iOjIsImV4cCI6MTcxNjI2MDU1NywiZGVwTmFtZSI6IuyduOyCrO2MgCJ9.NBWk8hm2pijWxZjy6Grk4PCIa3hzR2MY1unY3zKXVRY
 
 --WebAppBoundary
-Content-Disposition: form-data; name="mailInfo"
+Content-Disposition: form-data; name="mailDto"
 Content-Type: application/json
 
 {
